requirements.txt
requests==2.31.0
telebot==0.0.5
soundfile==0.12.1
openai==0.27.7
PyYAML==6.0b1
python-crontab==2.7.1
numpy==1.24.3